default the environment to the machine env
If `sfncli` encounters a workflow payload that lacks `_EXECUTION_NAME`,
the environment variables are not provided.

When not hitting the `if executionName != nil` codepath, `t.execCmd.Env`
would default to nil and then execute the process with either:

1. no environment variables
2. only the `WORK_DIR` env var

------

I think this should maybe be paired with a change to `workflow-manager`
to re-insert `_EXECUTION_NAME` to payloads whenever it is dropped by a
worker. Currently this is done by `sfncli`, but it relies on the value
being present in the input.